# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/activities": {
      "get": {
        "operationId": "FlowController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityListDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [],
        "summary": "",
        "tags": [
          "最小业务链路"
        ]
      },
      "post": {
        "operationId": "OrganizerController_create",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityWriteInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedActivityDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizer"
        ]
      }
    },
...
    "/api/v1/media/covers": {
      "post": {
        "operationId": "OrganizerController_uploadCover",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverUploadDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizer"
        ]
      }
    },
    "/api/v1/media/covers/{name}": {
      "get": {
        "operationId": "OrganizerController_coverImage",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [],
        "summary": "",
        "tags": [
          "organizer"
        ]
      }
    },
    "/api/v1/activities/{id}/edit": {
      "post": {
        "operationId": "OrganizerController_edit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityWriteInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedActivityDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizer"
        ]
      }
    },
    "/api/v1/activities/{id}/publish": {
      "post": {
        "operationId": "OrganizerController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedActivityDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizer"
        ]
      }
    },
    "/api/v1/members/me/activities": {
      "get": {
        "operationId": "OrganizerController_list",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedActivityListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizer"
        ]
      }
    },
    "/api/v1/activities/{id}/manage": {
      "get": {
        "operationId": "OrganizerController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedActivityDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizer"
        ]
      }
    },
    "/api/v1/activities/{id}/registrations/manage": {
      "get": {
        "operationId": "OrganizerController_roster",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizerRosterDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizer"
        ]
      }
    },
    "/api/v1/activities/{id}/cancel": {
      "post": {
        "operationId": "OrganizerController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelActivityInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedActivityDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizer"
        ]
      }
    },
    "/api/v1/members/me/notifications": {
      "get": {
        "operationId": "OrganizerController_notifications",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizer"
        ]
      }
    },
    "/api/v1/members/me/notifications/{id}/read": {
      "post": {
        "operationId": "OrganizerController_read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizer"
        ]
      }
    },
...
  },
  "info": {
    "title": "会聚 API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
...
      "OkDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "AvatarInput": {
        "type": "object",
        "properties": {
          "mimeType": {
            "type": "string",
            "enum": [
              "image/png",
              "image/jpeg"
            ]
          },
          "base64": {
            "type": "string"
          }
        },
        "required": [
          "mimeType",
          "base64"
        ]
      },
      "PublicMemberDto": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "memberId",
          "avatarUrl",
          "displayName"
        ]
      },
      "QuestionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "short_text",
              "long_text",
              "single",
              "multiple"
            ]
          },
          "prompt": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "options": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "type",
          "prompt",
          "required",
          "options"
        ]
      },
      "ActivityDto": {
        "type": "object",
        "properties": {
          "organizer": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PublicMemberDto"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "registrationDeadline": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "nullable": true
          },
          "activeRegistrationCount": {
            "type": "number"
          },
          "cancellationRegistrationCount": {
            "type": "number",
            "nullable": true
          },
          "feeType": {
            "type": "string",
            "enum": [
              "free",
              "paid"
            ]
          },
          "feeAmountCents": {
            "type": "number",
            "nullable": true
          },
          "registrationState": {
            "type": "string",
            "enum": [
              "open",
              "closed",
              "full",
              "cancelled",
              "removed"
            ]
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          },
          "consultationContact": {
            "type": "string"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionDto"
            }
          }
        },
        "required": [
          "organizer",
          "id",
          "title",
          "description",
          "coverUrl",
          "location",
          "startsAt",
          "endsAt",
          "registrationDeadline",
          "capacity",
          "activeRegistrationCount",
          "cancellationRegistrationCount",
          "feeType",
          "feeAmountCents",
          "registrationState",
          "cancellationReason",
          "questions"
        ]
      },
      "ActivityListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
      "AnswerInput": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string"
          },
          "value": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          }
        },
        "required": [
          "questionId",
          "value"
        ]
      },
...
      "CoverUploadDto": {
        "type": "object",
        "properties": {
          "coverUrl": {
            "type": "string"
          }
        },
        "required": [
          "coverUrl"
        ]
      },
      "QuestionWriteInput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "short_text",
              "long_text",
              "single",
              "multiple"
            ]
          },
          "prompt": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "options": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "prompt",
          "required"
        ]
      },
      "ActivityWriteInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string"
          },
          "consultationContact": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "registrationDeadline": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "number",
            "nullable": true
          },
          "feeType": {
            "type": "string",
            "enum": [
              "free",
              "paid"
            ]
          },
          "feeAmountCents": {
            "type": "number",
            "nullable": true
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionWriteInput"
            }
          }
        },
        "required": [
          "title",
          "description",
          "location",
          "consultationContact",
          "startsAt",
          "endsAt",
          "feeType",
          "questions"
        ]
      },
      "ManagedActivityDto": {
        "type": "object",
        "properties": {
          "organizer": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PublicMemberDto"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "registrationDeadline": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "nullable": true
          },
          "activeRegistrationCount": {
            "type": "number"
          },
          "cancellationRegistrationCount": {
            "type": "number",
            "nullable": true
          },
          "feeType": {
            "type": "string",
            "enum": [
              "free",
              "paid"
            ]
          },
          "feeAmountCents": {
            "type": "number",
            "nullable": true
          },
          "registrationState": {
            "type": "string",
            "enum": [
              "open",
              "closed",
              "full",
              "cancelled",
              "removed"
            ]
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          },
          "consultationContact": {
            "type": "string"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionDto"
            }
          },
          "lifecycle": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "cancelled"
            ]
          },
          "moderation": {
            "type": "string",
            "enum": [
              "normal",
              "removed"
            ]
          },
          "hasRegistrationEver": {
            "type": "boolean"
          }
        },
        "required": [
          "organizer",
          "id",
          "title",
          "description",
          "coverUrl",
          "location",
          "startsAt",
          "endsAt",
          "registrationDeadline",
          "capacity",
          "activeRegistrationCount",
          "cancellationRegistrationCount",
          "feeType",
          "feeAmountCents",
          "registrationState",
          "cancellationReason",
          "questions",
          "lifecycle",
          "moderation",
          "hasRegistrationEver"
        ]
      },
      "ManagedActivityListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManagedActivityDto"
            }
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "hasMore"
        ]
      },
      "OrganizerRegistrationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "activityId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled"
            ]
          },
          "contactPhone": {
            "type": "string"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerInput"
            }
          },
          "currentRegisteredAt": {
            "type": "string"
          },
          "cancelledAt": {
            "type": "string",
            "nullable": true
          },
          "member": {
            "$ref": "#/components/schemas/PublicMemberDto"
          },
          "attended": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "activityId",
          "status",
          "contactPhone",
          "answers",
          "currentRegisteredAt",
          "cancelledAt",
          "member",
          "attended"
        ]
      },
      "OrganizerRosterDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrganizerRegistrationDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CancelActivityInput": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ]
      },
      "NotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "activityId": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "activity_cancelled",
              "activity_removed",
              "activity_restored",
              "consultation_contact_updated"
            ]
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "readAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "activityId",
          "type",
          "title",
          "body",
          "readAt",
          "createdAt"
        ]
      },
      "NotificationListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationDto"
            }
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "hasMore"
        ]
      },
...
    }
  }
}